improved syntax error handling
