+++ src/parser/binary_tree/node_utils.c
@@ -6,7 +6,7 @@
 /*   By: padam <padam@student.42heilbronn.com>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/22 19:44:23 by padam             #+#    #+#             */
-/*   Updated: 2024/03/11 21:01:22 by padam            ###   ########.fr       */
+/*   Updated: 2024/03/13 15:04:02 by padam            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,9 +20,7 @@ t_node	*new_node(void)
 	if (!node)
 		return (NULL);
 	node->left = NULL;
-	node->new_process_left = false;
 	node->right = NULL;
-	node->new_process_right = false;
 	return (node);
 }
