+++ src/parser/cleanup.c
@@ -6,7 +6,7 @@
 /*   By: padam <padam@student.42heilbronn.com>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/26 23:10:31 by padam             #+#    #+#             */
-/*   Updated: 2024/03/12 17:13:50 by padam            ###   ########.fr       */
+/*   Updated: 2024/03/13 16:44:15 by padam            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
