+++ src/parser/parser.c
@@ -6,15 +6,11 @@
 /*   By: padam <padam@student.42heilbronn.com>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/15 21:15:56 by padam             #+#    #+#             */
-/*   Updated: 2024/03/12 14:58:49 by padam            ###   ########.fr       */
+/*   Updated: 2024/03/13 16:31:52 by padam            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "parser.h"
-
-void	debug_print_tree(t_node *node, int i);
-void	debug_print_redir(t_redir *redir, int i);
-void	debug_print_cmd(t_cmd *cmd);
 
 char	*get_current_folder(void)
 {
@@ -46,60 +42,6 @@ char	*new_prompt(void)
 	return (command);
 }
 
-void	debug_print_token_array(t_token *token_first)
-{
-	while (token_first)
-	{
-		printf("type: %d, value: %s\n", token_first->type, token_first->value);
-		token_first = token_first->next;
-	}
-}
-
-void	get_next_debug(void *ptr, t_node_type type, int i)
-{
-	i++;
-	if (type == CMD)
-		debug_print_cmd(ptr);
-	else if (type == ERROR)
-		printf("error\n");
-	else if (type == REDIR)
-		debug_print_redir(ptr, i);
-	else
-		debug_print_tree(ptr, i);
-}
-
-void	debug_print_cmd(t_cmd *cmd)
-{
-	int	i;
-
-	i = 0;
-	printf("----------------\n");
-	printf("args:\n");
-	while (cmd->args && cmd->args[i])
-		printf("%s\n", cmd->args[i++]);
-	i = 0;
-	printf("redirect_in: %d\n", cmd->redirect_in);
-	printf("redirect_out: %d\n", cmd->redirect_out);
-}
-
-void	debug_print_redir(t_redir *redir, int i)
-{
-	if (redir->redirects)
-		debug_print_cmd(redir->redirects);
-	if (redir->next)
-		get_next_debug(redir->next, redir->type, i);
-}
-
-void	debug_print_tree(t_node *node, int i)
-{
-	char *type_list[] = {"ERROR", "REDIR", "AND", "OR", "PIPE", "CMD", "SYNTAX"};
-	printf("%il%b: %s\n", i, node->new_process_left, type_list[node->type_left]);
-	printf("%ir%b: %s\n", i, node->new_process_right, type_list[node->type_right]);
-	if (node->left)
-		get_next_debug(node->left, node->type_left, i);
-	if (node->right)
-		get_next_debug(node->right, node->type_right, i);
-}
 
 void	parser(void)
 {
@@ -108,17 +50,22 @@ void	parser(void)
 	void	*token_tree;
 	t_node_type	token_tree_first;
 
+	command = NULL;
 	(void)token_tree_first;
 	while (1)
 	{
-		command = new_prompt();
+		while (!command || !*command)
+			command = new_prompt();
 		// command = "he && hi || du";
 		if (command)
 			tokens = tokenize_command(command);
 		free(command);
 		token_tree_first = tokens_to_tree(tokens, &token_tree);
+		if (token_tree_first == SYNTAX)
+			printf("syntax error\n");
 		if (token_tree)
 			get_next_debug(token_tree, token_tree_first, 0);
 		// debug_print_token_array(tokens);
+		command = NULL;
 	}
 }
